Fixes dmtxDecodeMatrixRegion returning a freed pointer
Call to dmtxDecodePopulatedArray can free msg argument,
which should no more be used when freed.

Unfortunately dmtxDecodePopulatedArray returned no hints
whereas it freed this pointer or not. This commit fixes
dmtxDecodePopulatedArray return type so that it informs
when it has freed its msg argument.

This commit fixes a regression inserted by e4233cac.